renamed kuzzleCollectionMapping.js -> CollectionMapping.js
